Issue#64 admin routes

+++ server/api/productsRoutes.js
@@ -27,8 +27,12 @@ router.get("/:productId", async (req, res, next) => {
 // POST/api/products   --------- Admin only
 router.post("/", async (req, res, next) => {
   try {
-    const newProduct = await Product.create(req.body);
-    res.status(201).json(newProduct);
+    if (req.user && req.user.isAdmin) {
+      const newProduct = await Product.create(req.body);
+      res.status(201).json(newProduct);
+    } else {
+      res.status(401).json({ error: "Unauthorized" });
+    }
   } catch (err) {
     next(err);
   }
@@ -37,16 +41,20 @@ router.post("/", async (req, res, next) => {
 // PUT/api/products   --------- Admin only
 router.put("/:productId", async (req, res, next) => {
   try {
-    for (let key in req.body) {
-      if (req.body[key] === "") {
-        delete req.body[key];
+    if (req.user && req.user.isAdmin) {
+      for (let key in req.body) {
+        if (req.body[key] === "") {
+          delete req.body[key];
+        }
       }
+      const product = await Product.findByPk(req.params.productId);
+
+      const editProduct = await product.update(req.body);
+
+      res.json(editProduct);
+    } else {
+      res.status(401).json({ error: "Unauthorized" });
     }
-    const product = await Product.findByPk(req.params.productId);
-
-    const editProduct = await product.update(req.body);
-
-    res.json(editProduct);
   } catch (err) {
     next(err);
   }
@@ -55,10 +63,14 @@ router.put("/:productId", async (req, res, next) => {
 // DELETE api/products   --------- Admin only
 router.delete("/:productId", async (req, res, next) => {
   try {
-    await Product.destroy({
-      where: { id: req.params.productId },
-    });
-    res.status(200).send(req.params.productId);
+    if (req.user && req.user.isAdmin) {
+      await Product.destroy({
+        where: { id: req.params.productId },
+      });
+      res.status(200).send(req.params.productId);
+    } else {
+      res.status(401).json({ error: "Unauthorized" });
+    }
   } catch (err) {
     next(err);
   }
